Update xmltoDataframe.R
Found a way to simplify the xml file to only the nodes we want using a perl script and xsl sxript before parsing in R. Once parsed in R, the xml file will produce a dataframe with precisely the data we want! I will post the perl and xsl scripts as well.