Found R command to call perl script
Not a working solution yet... I can have it call the script. But I don't
see the revised xml file in the folder. Need to find out why it doesn't
work as expected . I used absolute directory when calling the files
otherwise the R command to search for the file wouldn't work.