Update xmltoDataframe.R
Fixed a major issue with intersecting the matrices according to distance and latitude criteria. I realized that the bins weren't organized in the same way for each matrix so it was producing the wrong pairings but its fixed now. Added some features: bin size, minimum and maximum lat value determination plus a few minor tweaks to column names and presentation of data. Got rid of "BOLD:" in the bin_uri columns of dataframes. Added outgroup determination as well, still testing it, sometimes have a problem merging dfBestOutGroup to dfRandomBinSeq - seems to delete a row sometimes but other times works perfectly (I suspect my random bin generator code may be the issue). Added a finalized dataframe called dfMatchOverallBest which should have all of the important data Sally would need including pairings and outgroup. Much more detailed comments  in the script. In testing with reptilia I was getting between 10 and 20 pairings as a rough estimate of number of pairings you should be getting.