File upload
Added command to file upload the revised XML file to store in a variable
to be parsed instead of linking the absolute path. Allows better
functionality to the user instead of having the user to manually change
the pathway each time. System() to call the perl script doesn't need to
be saved into a variable, so removed it.