Create tsvtoDataFrame
Found a way to circumvent using perl altogether by using the tsv variant of files from the BOLD API. Can now grab all data from the BOLD API directly in tsv form and take that tsv and convert it into a dataframe in R. The program is very similar to xmltoDataframe after that intial step. I did also add export to csv and tsv.The user can also specify the name of the file being exported.