stats update
i managed to get the division of adjacent outgroup distances to work with monica's help. I have 2 for loops (1 for binomial because it adds the negative values to the division where the values where the higher-latitude lineage has the larger genetic distance to the outgroup and 1 for just wilcoxon where there doesn't need to be a negative value added, I think). The problem is that I cannot get the negative values to work and I tried a bunch of different things. But once that is fixed, everything should work. For now, I just have the histogram functions with btestOutGroup and trueTestOutGroupDist (binomial and wilcox respectively), not really sure what is supposed to be plotted.