Update server.R
Backing up the work I have. Integrated the backend script (tsvtoDataFrame.R) into this R script to connect with the user input. There is an error when changing the user input controls for some reason so need to fix that, the implementation of the markers on the map isn't finished neither is the plot for the statistical analysis. (may need to comment out the code otherwise it may not run properly at the moment)