Update tsvtoDataFrame.R
Near finalized script. Added sections relating to inputting and trimming based on the reference sequence. Also further expanded the identification of psuedoreplicates. Also fixed an issue where the ordering of dfAllSeq was different from the ordering after alignment, now dfAllSeq will be properly ordered based on the ordering of the alignment. Also added a section for removal of sequences containing high N content after the trimming stage. Need further clarification from Sally on the averaging of pseudoreplicates in the statistics section but thats the only outstanding item pending Sally's approval.